feat: finished refactor employee

+++ src/main/java/com/oocl/traning/Model/Employee.java
@@ -49,7 +49,7 @@ public class Employee {
         return salary;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
 }
